common: reorg proofs
